refac: scraper service
