+++ src/services/crons/cron-shop-0002.service.ts
@@ -1,14 +1,31 @@
 import { Injectable } from '@nestjs/common';
 import { Cron } from '@nestjs/schedule';
-import { ScraperShop0002 } from '../srapers/scraper-shop-0002.service';
 import { IDataForCron } from 'src/types/interfaces';
+import { ScraperUtilsService } from '../scraper.service';
+import { InjectRepository } from '@nestjs/typeorm';
+import { PricesShop0002 } from 'src/prices/prices-shop-0002.entity';
+import { Repository } from 'typeorm';
 
 @Injectable()
 export class CronShop0002 {
-  constructor(private readonly ScraperShop0002: ScraperShop0002) {}
+  constructor(
+    private readonly scraperUtilsService: ScraperUtilsService,
+
+    @InjectRepository(PricesShop0002)
+    private readonly prisesShopRepository: Repository<PricesShop0002>,
+  ) {}
+
+  parsePrice(price: string): number {
+    return Number(parseFloat(price.replace(/[^\d.]/g, '')).toFixed(2));
+    /*   const priceAsString = price.replace(/[^\d,]/g, '');
+    return parseFloat(
+      priceAsString.slice(0, -2) + '.' + priceAsString.slice(-2),
+    ); */
+  }
 
   //@Cron('0 1-23/2 * * *')
-  @Cron('15 * * * *')
+  //@Cron('*/15 * * * * *')
+  @Cron('43 * * * *')
   async handleCron() {
     const dataForCron: IDataForCron = {
       shop_id: 'shop-0002',
@@ -40,6 +57,11 @@ export class CronShop0002 {
         },
       ],
     };
-    await this.ScraperShop0002.scrape(dataForCron);
+    await this.scraperUtilsService.scrape(
+      dataForCron,
+      PricesShop0002,
+      this.prisesShopRepository,
+      this.parsePrice,
+    );
   }
 }
